+++ .speck/scripts/v7/regenerate-project-readme.sh
@@ -1,0 +1,1 @@
+../regenerate-project-readme.sh